+++ statsd/otel_cache.go
@@ -1,4 +1,4 @@
-package otel
+package statsd
 
 import (
 	"sync"
