+++ statsd/otel_format_tags.go
@@ -1,4 +1,4 @@
-package otel
+package statsd
 
 import (
 	"fmt"
